+++ src/api/Client.cpp
@@ -90,7 +90,8 @@ Client::live() const
 }
 
 Status
-Client::open(const ApplicationLayerProtocol alpn, const uint8_t options, ID& id)
+Client::open(const ApplicationLayerProtocol alpn, const uint16_t options,
+             ID& id)
 {
   /*
    * Open the connection from the TCP end.
